Reset version back to 0.3
I know this may cause some headaches, but we want to keep the version #
below 1.0 under we're fully in production. The 1.0 hack was a
workaround for pbr problems.

And we need to bump anyway for the "count" option on find_streams to
get exposed.

This is all really just for pypi.

Change-Id: I8d2567d303c04a5b9e6c295522568cca301283ec