Add mark/limit paging to find call.
Adds Mark/limit to find_streams for api usage.

The 'mark' is an opaque parameter (called '_mark') added into
each item in the list of returned items (currently a mildly obscured
ID for streams, that could be changed.)

Returned items are sorted in descending order (by ID for streams),
allowing paging from the 'top' to the bottom.
By prefacing the 'mark' param with a '+', you can page in ascending
order as well.

This same paging scheme can be used for a future events api, as well.

Change-Id: Ie358c0d1d8644dac24cad3992188ec22d66c5328