Only consider the last interesting event when checking traits.
Was previously looking at the last .start event when
comparing important traits against the .exists event.
Now it will look at the last important .end event it found (if any).

For delete operations, it will be the last event in the chain.
This may need to change to a specific event as well, like
compute.instance.delete.end or something.

Change-Id: Ie85e69716cd730c3546d06bae28e52e0c301940f